// src/types/theme.ts
// Define theme types
export type ThemeColors = {
  filesHeaderBgColor: string;
  mainContentBgColor: string;
  topHeaderBgColor: string;
  iconsSidebarBgColor: string;
  headerFileBgColor: string;
  explorerSidebarBgColor: string;

  activeLineColor: string;
  activeTabColor: string;

  activeColor: string;
  subTextColor: string;
  secondaryColor: string;

  lightGray: string;
  darkGray: string;
};

export type ThemeName =
  | "Github Dark"
  | "Solarized Light"
  | "Cobalt2"
  | "Github Light"
  | "Ayu Green Dark"
  | "Night Owl";